Solution to Convert Decimal to Binary

+++ Convert Decimal to Binary@@67882a731b4392d8fc1577f5/code.c
@@ -1,15 +1,15 @@
 // Your code here...
 #include <stdio.h>
 int main(){
-    int n;
+    long long n;
     scanf("%d",&n);
-    int rem, i =1;
-    int result = 0;
+    int  rem, i =1;
+    long long result = 0;
     while(n!=0){
         rem = n%2;
         n/=2;
         result += rem*i;
         i*10;
     }
-    printf("%d",result);
+    printf("%lld",result);
 }
